Extract `PreviewHelper` to it own file
